Part 2 III IV V

+++ AM.ApplicationCore/Domain/Flight.cs
@@ -12,7 +12,7 @@ namespace AM.ApplicationCore.Domain
         {
         }
 
-        public Flight(DateTime effectiveArrival, string departure, string destination, int estimatedDuration, DateTime flightDate, int flightId, Plane plane, ICollection<Passenger> passengers)
+        public Flight(DateTime effectiveArrival, string departure, string destination, double estimatedDuration, DateTime flightDate, int flightId, Plane plane, ICollection<Passenger> passengers)
         {
             EffectiveArrival = effectiveArrival;
             Departure = departure;
@@ -27,7 +27,7 @@ namespace AM.ApplicationCore.Domain
         public DateTime EffectiveArrival { get; set; }
         public string Departure { get; set; }
         public string Destination { get; set; }
-        public int EstimatedDuration { get; set; }
+        public double EstimatedDuration { get; set; }
         public DateTime FlightDate { get; set; }
         public int FlightId { get; set; }
         public Plane Plane { get; set; }
